SQL Script and Migration New

+++ database/migrations/2023_01_04_171643_create_outlet_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string("name");
             $table->string("picture")->nullable();
             $table->string("address")->nullable();
-            $table->double('latitude', 15,8)->nullable();
-            $table->double('longitude', 15,8)->nullable();
+            $table->decimal('latitude',20,8)->nullable();
+            $table->decimal('longitude',20,8)->nullable();
             $table->unsignedBigInteger("brand_id");
             $table->timestamps();
             $table->foreign('brand_id')->references('id')->on('brands');
